FIX: No border on focused links
This overrides the default stylesheet on firefox and chrome to not show a border around the currently focused element in the navigation drawer.
While it's not a problem with normal links, when opening the drawer the first link is focused and the border would be shown, which I don't want.

fixes #8