FEATURE: Persist the drawer on large screens
On large screens the drawer is displayed fixed on the left and the menubar
on the top is removed completely.
The action updateDrawerState is therefore expanded by an object parameter
which contains the properties opened and persistent.
All calls to this action are updated to only change the respective property.
Could probably also represent this as an object in the store instead of
two discrete properties :see_no_evil:

Also, the updateLayout action is removed as it was not needed.